feat(database): extract table truncation utility and isolate connection pooling
- Extract truncateAllTables and clearSchemaTablesCache from database.connection.ts to truncate.util.ts
- Guard truncateAllTables against production execution using validated env singleton
- Add test-env.ts preloaded via bunfig.toml to establish test environment before module evaluation
- Re-export truncation functions in database.helper.ts and update seed/test callers
